huajiyu 发布系统代码 [后台服务体系，iBiz软件生产管理]

+++ app_Web/src/widgets/product-stats/product-input-table-grid/product-input-table-grid-model.ts
@@ -39,7 +39,7 @@ export default class ProductInputTableModel {
         {
           name: 'haveconsumed',
           prop: 'haveconsumed',
-          dataType: 'FLOAT',
+          dataType: 'INT',
         },
         {
           name: 'srfmajortext',
